Rewrite NobugLandingOrigin to not have NaN origins
NaN origins will now mostly be replaced with direct traces to the ground. This is better than NaN's because it won't invalidate jumps for players. It also won't ever be longer than a normal jump that didn't get NaN'd. And more: it reduces the work that this api's end-users have to do to deal with NaN origins. I also changed some logic, so that sv_gravity doesn't need to be used anymore.